Revert "Extract the tool execution away from the runtime"
This reverts commit d73b5b5d20d64ed16df248dee8dd2d5da433dfa0.
